Styles now load synchronously
